wip: OpenApi Specs for all Controllers
-Updated OpenApi spec for all controllers
