+++ api/Controllers/v1/Edition.php
@@ -17,62 +17,6 @@ use OpenApi\Attributes as OA;
  * logger - which returns an instance of \Monolog\Logger. This is also a protected property on your controller. Access it using $this->logger.
  */
 
-#[OA\OpenApi(
-    openapi: '3.1.0',
-    info: new OA\Info(
-        version: 'v1',
-        description: '<b />AlQuran API - Edition',
-        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
-    ),
-    servers: [
-        new OA\Server(url: 'https://api.alquran.cloud/v1'),
-        new OA\Server(url: 'http://api.alquran.cloud/v1')
-    ],
-    tags: [
-        new OA\Tag(name: 'Edition')
-    ]
-)]
-#[OA\Components(
-    schemas: [
-        new OA\Schema(
-            schema: '200EditionQuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: 'quran-uthmani-quran-academy'),
-                new OA\Property(property: 'language', type: 'string', example: 'ar'),
-                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم برسم العثماني (quran-academy)" ),
-                new OA\Property(property: 'englishName', type: 'string', example: "Modified Quran Uthmani Text from the Quran Academy to work with the Kitab font"),
-                new OA\Property(property: 'format', type: 'string', example: 'text'),
-                new OA\Property(property: 'type', type: 'string', example: 'quran'),
-                new OA\Property(property: 'direction', type: 'string', example: 'rtl')
-            ]
-        ),
-        new OA\Schema(
-            schema: '200EditionQuranSimpleResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: 'quran-simple'),
-                new OA\Property(property: 'language', type: 'string', example: 'ar'),
-                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم المبسط (تشكيل بسيط) (simple)"),
-                new OA\Property(property: 'englishName', type: 'string', example: "Simple"),
-                new OA\Property(property: 'format', type: 'string', example: 'text'),
-                new OA\Property(property: 'type', type: 'string', example: 'quran'),
-                new OA\Property(property: 'direction', type: 'string', example: 'rtl')
-            ]
-        )
-    ],
-    responses: [
-        new OA\Response(response: '404EditionResponse', description: 'Unable to find the requested resource',
-            content: new OA\MediaType(mediaType: 'application/json',
-                schema: new OA\Schema(
-                    properties: [
-                        new OA\Property(property: 'code', type: 'integer', example: 404),
-                        new OA\Property(property: 'status', type: 'string', example: 'RESOURCE_NOT_FOUND'),
-                        new OA\Property(property: 'data', type: 'string', example: 'Not found.')
-                    ]
-                )
-            )
-        )
-    ]
-)]
 
 class Edition extends AlQuranController
 {
@@ -109,7 +53,7 @@ class Edition extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404EditionResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
 
@@ -165,7 +109,7 @@ class Edition extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404EditionResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
 
@@ -270,7 +214,7 @@ class Edition extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404EditionResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
 
@@ -372,7 +316,7 @@ class Edition extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404EditionResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
