+++ api/Controllers/v1/HizbQuarter.php
@@ -17,97 +17,6 @@ use OpenApi\Attributes as OA;
  * logger - which returns an instance of \Monolog\Logger. This is also a protected property on your controller. Access it using $this->logger.
  */
 
-#[OA\OpenApi(
-    openapi: '3.1.0',
-    info: new OA\Info(
-        version: 'v1',
-        description: '<p><b />AlQuran API - HizbQuarter</p>
-        The Holy Quran comprises 240 Hizb Quarters. One Hizb is half a Juz.',
-        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
-    ),
-    servers: [
-        new OA\Server(url: 'https://api.alquran.cloud/v1'),
-        new OA\Server(url: 'http://api.alquran.cloud/v1')
-    ],
-    tags: [
-        new OA\Tag(name: 'HizbQuarter')
-    ]
-)]
-#[OA\Components(
-    schemas: [
-        new OA\Schema(
-            schema: '200EditionQuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: 'quran-uthmani-quran-academy'),
-                new OA\Property(property: 'language', type: 'string', example: 'ar'),
-                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم برسم العثماني (quran-academy)" ),
-                new OA\Property(property: 'englishName', type: 'string', example: "Modified Quran Uthmani Text from the Quran Academy to work with the Kitab font"),
-                new OA\Property(property: 'format', type: 'string', example: 'text'),
-                new OA\Property(property: 'type', type: 'string', example: 'quran'),
-                new OA\Property(property: 'direction', type: 'string', example: 'rtl')
-            ]
-        ),
-        new OA\Schema(
-            schema: '200HizbQuarterQuranUthmaniPartialSurahResponse',
-            properties: [
-                new OA\Property(property: 'number', type: 'integer', example: 1),
-                new OA\Property(property: 'name', type: 'string', example: "سُورَةُ ٱلْفَاتِحَةِ"),
-                new OA\Property(property: 'englishName', type: 'string', example: "Al-Faatiha"),
-                new OA\Property(property: 'englishNameTranslation', type: 'string', example: "The Opening"),
-                new OA\Property(property: 'revelationType', type: 'string', example: 'Meccan'),
-                new OA\Property(property: 'numberOfAyahs', type: 'integer', example: 7)
-            ], type: 'object'
-        ),
-        new OA\Schema(
-            schema: '200HizbQuarterQuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'number', type: 'integer', example: 1),
-                new OA\Property(property: 'ayahs', type: 'array',
-                    items: new OA\Items(
-                        properties: [
-                            new OA\Property(property: 'number', type: 'integer', example: 5),
-                            new OA\Property(property: 'text', type: 'string', example: "إِیَّاكَ نَعۡبُدُ وَإِیَّاكَ نَسۡتَعِینُ"),
-                            new OA\Property(property: 'surah', ref: '#/components/schemas/200HizbQuarterQuranUthmaniPartialSurahResponse', type: 'object'),
-                            new OA\Property(property: 'numberInSurah', type: 'integer', example: 5),
-                            new OA\Property(property: 'juz', type: 'integer', example: 1),
-                            new OA\Property(property: 'manzil', type: 'integer', example: 1),
-                            new OA\Property(property: 'page', type: 'integer', example: 1),
-                            new OA\Property(property: 'ruku', type: 'integer', example: 1),
-                            new OA\Property(property: 'hizbQuarter', type: 'integer', example: 1),
-                            new OA\Property(property: 'sajda', type: 'boolean', example: false)
-                        ], type: 'object'
-                    )
-                ),
-                new OA\Property(property: 'surahs',
-                    properties: [
-                        new OA\Property(property: '1', ref: '#/components/schemas/200HizbQuarterQuranUthmaniPartialSurahResponse', type: 'object')
-                    ], type: 'object'
-                ),
-                new OA\Property(property: 'edition', ref: '#/components/schemas/200EditionQuranUthmaniResponse', type: 'object')
-            ], type: 'object'
-        )
-    ],
-    responses: [
-        new OA\Response(response: '404HizbQuarterResponse', description: 'Hizb Quarter - Not Found',content: new OA\MediaType(mediaType: 'application/json',
-            schema: new OA\Schema(
-                properties: [
-                    new OA\Property(property: 'code', type: 'integer', example: 404),
-                    new OA\Property(property: 'status', type: 'string', example: 'NOT FOUND'),
-                    new OA\Property(property: 'data', type: 'string', example: "HizbQuarter number should be between 1 and 240")
-                ]
-            )
-        ))
-    ],
-    parameters: [
-        new OA\PathParameter(parameter: 'HizbQuarterNumberParameter', name: 'number', description: 'Hizb Quarter Number',
-            in: 'path', required: true, schema: new OA\Schema(type: 'integer'), example: 1),
-        new OA\QueryParameter(parameter: 'HizbQuarterOffsetQueryParameter', name: 'offset', description: 'Offset ayahs in a Hizb Quarter by the given number',
-            in: 'query', required: false, schema: new OA\Schema(type: 'integer'), example: 4),
-        new OA\QueryParameter(parameter: 'HizbQuarterLimitQueryParameter', name: 'limit', description: 'This is the number of ayahs that the response will be limited to',
-            in: 'query', required: false, schema: new OA\Schema(type: 'integer'), example: 1),
-    ]
-)]
-
 class HizbQuarter extends AlQuranController
 {
 
@@ -119,7 +28,7 @@ class HizbQuarter extends AlQuranController
         parameters: [
             new OA\PathParameter(ref: '#/components/parameters/HizbQuarterNumberParameter'),
             new OA\QueryParameter(ref: '#/components/parameters/HizbQuarterOffsetQueryParameter'),
-            new OA\QueryParameter(ref: '#/components/parameters/HizbQuarterLimitQueryParameter')
+            new OA\QueryParameter(ref: '#/components/parameters/LimitQueryParameter')
         ],
         responses: [
             new OA\Response(response: '200', description: 'Returns the Hizb Quarter for the requested number',
@@ -128,7 +37,7 @@ class HizbQuarter extends AlQuranController
                         properties: [
                             new OA\Property(property: 'code', type: 'integer', example: 200),
                             new OA\Property(property: 'status', type: 'string', example: 'OK'),
-                            new OA\Property(property: 'data', ref: '#/components/schemas/200HizbQuarterQuranUthmaniResponse', type: 'object')
+                            new OA\Property(property: 'data', ref: '#/components/schemas/200QuranUthmaniEntireResponse', type: 'object')
                         ]
                     )
                 )
@@ -174,7 +83,7 @@ class HizbQuarter extends AlQuranController
             new OA\PathParameter(name: 'edition', description: 'Edition name',
                 in: 'path', required: true, schema: new OA\Schema(type: 'string'), example: 'quran-uthmani-quran-academy'),
             new OA\QueryParameter(ref: '#/components/parameters/HizbQuarterOffsetQueryParameter'),
-            new OA\QueryParameter(ref: '#/components/parameters/HizbQuarterLimitQueryParameter')
+            new OA\QueryParameter(ref: '#/components/parameters/LimitQueryParameter')
         ],
         responses: [
             new OA\Response(response: '200', description: 'Returns the Hizb Quarter for the requested number and requested edition',
@@ -183,7 +92,7 @@ class HizbQuarter extends AlQuranController
                         properties: [
                             new OA\Property(property: 'code', type: 'integer', example: 200),
                             new OA\Property(property: 'status', type: 'string', example: 'OK'),
-                            new OA\Property(property: 'data', ref: '#/components/schemas/200HizbQuarterQuranUthmaniResponse', type: 'object')
+                            new OA\Property(property: 'data', ref: '#/components/schemas/200QuranUthmaniEntireResponse', type: 'object')
                         ]
                     )
                 )
