+++ api/Controllers/v1/Quran.php
@@ -22,83 +22,83 @@ use OpenApi\Attributes as OA;
  * logger - which returns an instance of \Monolog\Logger. This is also a protected property on your controller. Access it using $this->logger.
  */
 
-#[OA\OpenApi(
-    openapi: '3.1.0',
-    info: new OA\Info(
-        version: 'v1',
-        description: '<p><b />AlQuran API - Quran
-        <br />The Holy Quran is a divine book from Allah SWT.</p>
-        You can get the text for complete edition using the endpoints below.',
-        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
-    ),
-    servers: [
-        new OA\Server(url: 'https://api.alquran.cloud/v1'),
-        new OA\Server(url: 'http://api.alquran.cloud/v1')
-    ],
-    tags: [
-        new OA\Tag(name: 'Quran')
-    ]
-)]
-#[OA\Components(
-    schemas: [
-        new OA\Schema(
-            schema: '200QuranUthmaniEditionResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: 'quran-uthmani-quran-academy'),
-                new OA\Property(property: 'language', type: 'string', example: 'ar'),
-                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم برسم العثماني (quran-academy)" ),
-                new OA\Property(property: 'englishName', type: 'string', example: "Modified Quran Uthmani Text from the Quran Academy to work with the Kitab font"),
-                new OA\Property(property: 'format', type: 'string', example: 'text'),
-                new OA\Property(property: 'type', type: 'string', example: 'quran')
-            ]
-        ),
-        new OA\Schema(
-            schema: '200QuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'surahs', type: 'array',
-                    items: new OA\Items(
-                        properties: [
-                            new OA\Property(property: 'number', type: 'integer', example: 1),
-                            new OA\Property(property: 'name', type: 'string', example: "سُورَةُ ٱلْفَاتِحَةِ"),
-                            new OA\Property(property: 'englishName', type: 'string', example: "Al-Faatiha"),
-                            new OA\Property(property: 'englishNameTranslation', type: 'string', example: "The Opening"),
-                            new OA\Property(property: 'revelationType', type: 'string', example: 'Meccan'),
-                            new OA\Property(property: 'ayahs', type: 'array',
-                                items: new OA\Items(
-                                    properties: [
-                                        new OA\Property(property: 'number', type: 'integer', example: 1),
-                                        new OA\Property(property: 'text', type: 'string', example: "بِسۡمِ ٱللَّهِ ٱلرَّحۡمَـٰنِ ٱلرَّحِیمِ"),
-                                        new OA\Property(property: 'numberInSurah', type: 'integer', example: 1),
-                                        new OA\Property(property: 'juz', type: 'integer', example: 1),
-                                        new OA\Property(property: 'manzil', type: 'integer', example: 1),
-                                        new OA\Property(property: 'page', type: 'integer', example: 1),
-                                        new OA\Property(property: 'ruku', type: 'integer', example: 1),
-                                        new OA\Property(property: 'hizbQuarter', type: 'integer', example: 1),
-                                        new OA\Property(property: 'sajda', type: 'boolean', example: false)
-                                    ], type: 'object'
-                                )
-                            )
-                        ], type: 'object'
-                    )
-                ),
-                new OA\Property(property: 'edition', ref: '#/components/schemas/200QuranUthmaniEditionResponse', type: 'object')
-            ], type: 'object'
-        )
-    ],
-    responses: [
-        new OA\Response(response: '404QuranResponse', description: 'Unable to find the requested resource',
-            content: new OA\MediaType(mediaType: 'application/json',
-                schema: new OA\Schema(
-                    properties: [
-                        new OA\Property(property: 'code', type: 'integer', example: 404),
-                        new OA\Property(property: 'status', type: 'string', example: 'RESOURCE_NOT_FOUND'),
-                        new OA\Property(property: 'data', type: 'string', example: 'Not found.')
-                    ]
-                )
-            )
-        )
-    ]
-)]
+//#[OA\OpenApi(
+//    openapi: '3.1.0',
+//    info: new OA\Info(
+//        version: 'v1',
+//        description: '<p><b />AlQuran API - Quran
+//        <br />The Holy Quran is a divine book from Allah SWT.</p>
+//        You can get the text for complete edition using the endpoints below.',
+//        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
+//    ),
+//    servers: [
+//        new OA\Server(url: 'https://api.alquran.cloud/v1'),
+//        new OA\Server(url: 'http://api.alquran.cloud/v1')
+//    ],
+//    tags: [
+//        new OA\Tag(name: 'Quran')
+//    ]
+//)]
+//#[OA\Components(
+//    schemas: [
+//        new OA\Schema(
+//            schema: '200QuranUthmaniEditionResponse',
+//            properties: [
+//                new OA\Property(property: 'identifier', type: 'string', example: 'quran-uthmani-quran-academy'),
+//                new OA\Property(property: 'language', type: 'string', example: 'ar'),
+//                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم برسم العثماني (quran-academy)" ),
+//                new OA\Property(property: 'englishName', type: 'string', example: "Modified Quran Uthmani Text from the Quran Academy to work with the Kitab font"),
+//                new OA\Property(property: 'format', type: 'string', example: 'text'),
+//                new OA\Property(property: 'type', type: 'string', example: 'quran')
+//            ]
+//        ),
+//        new OA\Schema(
+//            schema: '200QuranUthmaniResponse',
+//            properties: [
+//                new OA\Property(property: 'surahs', type: 'array',
+//                    items: new OA\Items(
+//                        properties: [
+//                            new OA\Property(property: 'number', type: 'integer', example: 1),
+//                            new OA\Property(property: 'name', type: 'string', example: "سُورَةُ ٱلْفَاتِحَةِ"),
+//                            new OA\Property(property: 'englishName', type: 'string', example: "Al-Faatiha"),
+//                            new OA\Property(property: 'englishNameTranslation', type: 'string', example: "The Opening"),
+//                            new OA\Property(property: 'revelationType', type: 'string', example: 'Meccan'),
+//                            new OA\Property(property: 'ayahs', type: 'array',
+//                                items: new OA\Items(
+//                                    properties: [
+//                                        new OA\Property(property: 'number', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'text', type: 'string', example: "بِسۡمِ ٱللَّهِ ٱلرَّحۡمَـٰنِ ٱلرَّحِیمِ"),
+//                                        new OA\Property(property: 'numberInSurah', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'juz', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'manzil', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'page', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'ruku', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'hizbQuarter', type: 'integer', example: 1),
+//                                        new OA\Property(property: 'sajda', type: 'boolean', example: false)
+//                                    ], type: 'object'
+//                                )
+//                            )
+//                        ], type: 'object'
+//                    )
+//                ),
+//                new OA\Property(property: 'edition', ref: '#/components/schemas/200QuranUthmaniEditionResponse', type: 'object')
+//            ], type: 'object'
+//        )
+//    ],
+//    responses: [
+//        new OA\Response(response: '404QuranResponse', description: 'Unable to find the requested resource',
+//            content: new OA\MediaType(mediaType: 'application/json',
+//                schema: new OA\Schema(
+//                    properties: [
+//                        new OA\Property(property: 'code', type: 'integer', example: 404),
+//                        new OA\Property(property: 'status', type: 'string', example: 'RESOURCE_NOT_FOUND'),
+//                        new OA\Property(property: 'data', type: 'string', example: 'Not found.')
+//                    ]
+//                )
+//            )
+//        )
+//    ]
+//)]
 
 class Quran extends AlQuranController
 {
@@ -120,7 +120,7 @@ class Quran extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404QuranResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
 
@@ -169,7 +169,7 @@ class Quran extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404QuranResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
