+++ api/Controllers/v1/Sajda.php
@@ -17,90 +17,6 @@ use OpenApi\Attributes as OA;
  * logger - which returns an instance of \Monolog\Logger. This is also a protected property on your controller. Access it using $this->logger.
  */
 
-#[OA\OpenApi(
-    openapi: '3.1.0',
-    info: new OA\Info(
-        version: 'v1',
-        description: '<p><b />AlQuran API - Sajda</p>
-        You can get all verses requiring Sajda / Prostration in the Holy Quran using the endpoints below.',
-        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
-    ),
-    servers: [
-        new OA\Server(url: 'https://api.alquran.cloud/v1'),
-        new OA\Server(url: 'http://api.alquran.cloud/v1')
-    ],
-    tags: [
-        new OA\Tag(name: 'Sajda')
-    ]
-)]
-#[OA\Components(
-    schemas: [
-        new OA\Schema(
-            schema: '200SajdaEditionQuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: 'quran-uthmani-quran-academy'),
-                new OA\Property(property: 'language', type: 'string', example: 'ar'),
-                new OA\Property(property: 'name', type: 'string', example: "القرآن الكريم برسم العثماني (quran-academy)" ),
-                new OA\Property(property: 'englishName', type: 'string', example: "Modified Quran Uthmani Text from the Quran Academy to work with the Kitab font"),
-                new OA\Property(property: 'format', type: 'string', example: 'text'),
-                new OA\Property(property: 'type', type: 'string', example: 'quran'),
-                new OA\Property(property: 'direction', type: 'string', example: 'rtl')
-            ]
-        ),
-        new OA\Schema(
-            schema: '200SajdaQuranUthmaniPartialSurahResponse',
-            properties: [
-                new OA\Property(property: 'number', type: 'integer', example: 7),
-                new OA\Property(property: 'name', type: 'string', example:  "سُورَةُ الأَعۡرَافِ"),
-                new OA\Property(property: 'englishName', type: 'string', example: "Al-A'raaf"),
-                new OA\Property(property: 'englishNameTranslation', type: 'string', example: "The Heights"),
-                new OA\Property(property: 'revelationType', type: 'string', example: 'Meccan'),
-                new OA\Property(property: 'numberOfAyahs', type: 'integer', example: 206)
-            ], type: 'object'
-        ),
-        new OA\Schema(
-            schema: '200SajdaQuranUthmaniResponse',
-            properties: [
-                new OA\Property(property: 'ayahs', type: 'array',
-                    items: new OA\Items(
-                        properties: [
-                            new OA\Property(property: 'number', type: 'integer', example: 1160),
-                            new OA\Property(property: 'text', type: 'string', example: "إِنَّ ٱلَّذِینَ عِندَ رَبِّكَ لَا یَسۡتَكۡبِرُونَ عَنۡ عِبَادَتِهِۦ وَیُسَبِّحُونَهُۥ وَلَهُۥ یَسۡجُدُونَ ۩"),
-                            new OA\Property(property: 'surah', ref: '#/components/schemas/200SajdaQuranUthmaniPartialSurahResponse', type: 'object'),
-                            new OA\Property(property: 'numberInSurah', type: 'integer', example: 206),
-                            new OA\Property(property: 'juz', type: 'integer', example: 9),
-                            new OA\Property(property: 'manzil', type: 'integer', example: 2),
-                            new OA\Property(property: 'page', type: 'integer', example: 176),
-                            new OA\Property(property: 'ruku', type: 'integer', example: 145),
-                            new OA\Property(property: 'hizbQuarter', type: 'integer', example: 70),
-                            new OA\Property(property: 'sajda',
-                                properties: [
-                                    new OA\Property(property: 'id', type: 'integer', example: 1),
-                                    new OA\Property(property: 'recommended', type: 'boolean', example: true),
-                                    new OA\Property(property: 'obligatory', type: 'boolean', example: false)
-                                ], type: 'object'
-                            )
-                        ], type: 'object'
-                    )
-                ),
-                new OA\Property(property: 'edition', ref: '#/components/schemas/200SajdaEditionQuranUthmaniResponse', type: 'object')
-            ], type: 'object'
-        )
-    ],
-    responses: [
-        new OA\Response(response: '404SajdaResponse', description: 'Unable to find the requested resource',
-            content: new OA\MediaType(mediaType: 'application/json',
-                schema: new OA\Schema(
-                    properties: [
-                        new OA\Property(property: 'code', type: 'integer', example: 404),
-                        new OA\Property(property: 'status', type: 'string', example: 'RESOURCE_NOT_FOUND'),
-                        new OA\Property(property: 'data', type: 'string', example: 'Not found.')
-                    ]
-                )
-            )
-        )
-    ]
-)]
 
 
 class Sajda extends AlQuranController
@@ -123,7 +39,7 @@ class Sajda extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404SajdaResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
 
@@ -172,7 +88,7 @@ class Sajda extends AlQuranController
                     )
                 )
             ),
-            new OA\Response(ref: '#/components/responses/404SajdaResponse', response: '404')
+            new OA\Response(ref: '#/components/responses/404NotFoundResourceResponse', response: '404')
         ]
     )]
