+++ api/Controllers/v1/Search.php
@@ -17,80 +17,6 @@ use OpenApi\Attributes as OA;
  * logger - which returns an instance of \Monolog\Logger. This is also a protected property on your controller. Access it using $this->logger.
  */
 
-#[OA\OpenApi(
-    openapi: '3.1.0',
-    info: new OA\Info(
-        version: 'v1',
-        description: '<p><b />AlQuran API - Search</p>
-        Search the text of the Holy Quran. Please note that only text editions of the Holy Quran are searchable.',
-        title: 'بِسْمِ اللهِ الرَّحْمٰنِ الرَّحِيْمِ'
-    ),
-    servers: [
-        new OA\Server(url: 'https://api.alquran.cloud/v1'),
-        new OA\Server(url: 'http://api.alquran.cloud/v1')
-    ],
-    tags: [
-        new OA\Tag(name: 'Search')
-    ]
-)]
-#[OA\Components(
-    schemas: [
-        new OA\Schema(
-            schema: '200SearchEditionSahihResponse',
-            properties: [
-                new OA\Property(property: 'identifier', type: 'string', example: "en.sahih"),
-                new OA\Property(property: 'language', type: 'string', example: 'en'),
-                new OA\Property(property: 'name', type: 'string', example: "Saheeh International"),
-                new OA\Property(property: 'englishName', type: 'string', example: "Saheeh International"),
-                new OA\Property(property: 'type', type: 'string', example: 'translation')
-            ]
-        ),
-        new OA\Schema(
-            schema: '200SearchPartialSurahResponse',
-            properties: [
-                new OA\Property(property: 'number', type: 'integer', example: 1),
-                new OA\Property(property: 'name', type: 'string', example: "سُورَةُ ٱلْفَاتِحَةِ"),
-                new OA\Property(property: 'englishName', type: 'string', example: "Al-Faatiha"),
-                new OA\Property(property: 'englishNameTranslation', type: 'string', example: "The Opening"),
-                new OA\Property(property: 'revelationType', type: 'string', example: 'Meccan')
-            ], type: 'object'
-        ),
-        new OA\Schema(
-            schema: '200SearchSahihResponse',
-            properties: [
-                new OA\Property(property: 'count', type: 'integer', example: 25),
-                new OA\Property(property: 'matches', type: 'array',
-                    items: new OA\Items(
-                        properties: [
-                            new OA\Property(property: 'number', type: 'integer', example: 6),
-                            new OA\Property(property: 'text', type: 'string', example: "Guide us to the straight path -"),
-                            new OA\Property(property: 'edition', ref: '#/components/schemas/200SearchEditionSahihResponse', type: 'object'),
-                            new OA\Property(property: 'surah', ref: '#/components/schemas/200SearchPartialSurahResponse', type: 'object'),
-                            new OA\Property(property: 'numberInSurah', type: 'integer', example: 6),
-                        ], type: 'object'
-                    )
-                )
-            ], type: 'object'
-        )
-    ],
-    responses: [
-        new OA\Response(response: '404SearchResponse', description: "User's Search - Not Found",content: new OA\MediaType(mediaType: 'application/json',
-            schema: new OA\Schema(
-                properties: [
-                    new OA\Property(property: 'code', type: 'integer', example: 404),
-                    new OA\Property(property: 'status', type: 'string', example: 'NOT FOUND'),
-                    new OA\Property(property: 'data', type: 'string', example: "Nothing matching your search was found..")
-                ]
-            )
-        ))
-    ],
-    parameters: [
-        new OA\PathParameter(parameter: 'SearchWordParameter', name: 'word', description: 'Word to search in text',
-            in: 'path', required: true, schema: new OA\Schema(type: 'string'), example: 'Path'),
-        new OA\PathParameter(parameter: 'SearchSurahParameter', name: 'surah', description: "Enter a surah number (between 1 and 114) to search a specific surah or 'all' to search all the text",
-            in: 'path', required: true, schema: new OA\Schema(type: 'string'), example: '1'),
-    ]
-)]
 
 class Search extends AlQuranController
 {
